+++ src/common/db/Database.py
@@ -460,7 +460,7 @@ class Database:
                 metadata = session.query(Metadata).with_entities(Metadata.version).filter_by(id=1).first()
                 if metadata:
                     return metadata.version
-                return "1.6.3-rc3"
+                return "1.6.3"
             except BaseException as e:
                 return f"Error: {e}"
 
@@ -492,7 +492,7 @@ class Database:
             "last_instances_change": None,
             "reload_ui_plugins": False,
             "integration": "unknown",
-            "version": "1.6.3-rc3",
+            "version": "1.6.3",
             "database_version": "Unknown",  # ? Extracted from the database
             "default": True,  # ? Extra field to know if the returned data is the default one
         }
@@ -2182,7 +2182,7 @@ class Database:
                             break
 
                 if matched_group is not None:
-                    multiple.setdefault(matched_group, {}).setdefault(window, set()).add(match.group("suffix"))
+                    multiple.setdefault(matched_group, {}).setdefault(window, set()).add(int(match.group("suffix")))
 
             if new_value is not None:
                 config[key] = new_value
@@ -2198,7 +2198,7 @@ class Database:
 
                     for window, suffixes in multiple[group_key].items():
                         template = templates.get(window, "") or templates.get("global", "")
-                        for suffix in suffixes:
+                        for suffix in map(int, suffixes):
                             if window == "global" or service:
                                 key = f"{setting.id}_{suffix}"
                             else: