+++ src/ui/app/static/js/pages/reports.js
@@ -124,7 +124,7 @@ $(document).ready(function () {
   layout.topStart.buttons = [
     {
       extend: "colvis",
-      columns: "th:not(:nth-child(-n+3))",
+      columns: "th:not(:nth-child(-n+2))",
       text: `<span class="tf-icons bx bx-columns bx-18px me-md-2"></span><span class="d-none d-md-inline" data-i18n="button.columns">${t(
         "button.columns",
         "Columns",
@@ -235,7 +235,7 @@ $(document).ready(function () {
   const reports_config = {
     tableSelector: "#reports",
     tableName: "reports",
-    columnVisibilityCondition: (column) => column > 3 && column < 13,
+    columnVisibilityCondition: (column) => column > 2 && column < 13,
     dataTableOptions: {
       columnDefs: [
         { orderable: false, targets: -1 },