+++ src/ui/app/utils.py
@@ -74,6 +74,7 @@ COLUMNS_PREFERENCES_DEFAULTS = {
         "8": True,
     },
     "reports": {
+        "3": True,
         "4": True,
         "5": False,
         "6": True,