+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## v1.6.3-rc3 - ????/??/??
+## v1.6.3 - ????/??/??
+
+- [BUGFIX] Fix errors with `PostgreSQL` database, ensuring that suffixes are stored as integers for consistency.
+- [FEATURE] Enhance `Redirect` plugin to support multiple source/destination paths
+- [FEATURE] Enhance `Antibot` CAPTCHA functionality with customizable character set via the `ANTIBOT_CAPTCHA_ALPHABET` setting, allowing users to define a custom alphabet for CAPTCHA generation.
+- [UI] Always display the selected service and selected type when editing/creating a custom configuration
+- [UI] Add global configuration fetching functionality to easy mode
+- [LINUX] Add installation type to the post-install script to allow users to choose among `all-in-one` (Full installation), `manager` (Scheduler and UI), `worker` (BunkerWeb only), `scheduler` (Scheduler only), and `ui` (UI only) installation types.
+- [ALL-IN-ONE] In entrypoint script, create redis directory if it does not exist to avoid issues with Redis not starting properly.
+- [CONTRIBUTION] Thank you [Arakmar](https://github.com/Arakmar) for your contribution regarding the web UI's `reports` page.
+
+## v1.6.3-rc3 - 2025/07/30
 
 - [BUGFIX] Fix HTTP/3 not working on default server as the `reuseport` directive was missing in the `default-server-http.conf` file.
 - [UI] Fix missing settings when cloning a service in the web UI